common utilies are created and classification models' scripts adjusted
