+++ strategicClassification-HardtComparison.py
@@ -11,7 +11,7 @@ import math
 import os
 from datetime import datetime
 from src.strategic_classification.utils.gain_and_cost_func import score, f, g, f_derivative
-
+from src.strategic_classification.utils.data_utils import split_data, load_spam_dataset
 torch.set_default_dtype(torch.float64)
 torch.manual_seed(0)
 np.random.seed(0)
@@ -24,49 +24,6 @@ EPSILON = 0.3
 X_LOWER_BOUND = -10
 X_UPPER_BOUND = 10
 
-# Utils
-
-def split_data(X, Y, percentage):
-    num_val = int(len(X)*percentage)
-    return X[num_val:], Y[num_val:], X[:num_val], Y[:num_val]
-
-def shuffle(X, Y):
-    data = torch.cat((X, Y), 1)
-    data = data[torch.randperm(data.size()[0])]
-    X = data[:, :2]
-    Y = data[:, 2]
-    return X, Y
-
-def conf_mat(Y1, Y2):
-    num_of_samples = len(Y1)
-    mat = confusion_matrix(Y1, Y2, labels=[-1, 1])*100/num_of_samples
-    acc = np.trace(mat)
-    return mat, acc
-
-def calc_accuracy(Y, Ypred):
-    num = len(Y)
-    temp = Y - Ypred
-    acc = len(temp[temp == 0])*1./num
-    return acc
-
-# Dataset
-
-def load_spam_dataset():
-    torch.manual_seed(0)
-    np.random.seed(0)
-    path = r"C:\Users\sagil\Desktop\nir_project\tip_spam_data\IS_journal_tip_spam.arff"
-    data, meta = arff.loadarff(path)
-    df = pd.DataFrame(data)
-    most_disc = ['qTips_plc', 'rating_plc', 'qEmail_tip', 'qContacts_tip', 'qURL_tip', 'qPhone_tip', 'qNumeriChar_tip', 'sentistrength_tip', 'combined_tip', 'qWords_tip', 'followers_followees_gph', 'qunigram_avg_tip', 'qTips_usr', 'indeg_gph', 'qCapitalChar_tip', 'class1']
-    df = df[most_disc]
-    df["class1"].replace({b'spam': -1, b'notspam': 1}, inplace=True)
-    df = df.sample(frac=1).reset_index(drop=True)
-
-    Y = df['class1'].values
-    X = df.drop('class1', axis = 1).values
-    X -= np.mean(X, axis=0)
-    X /= np.std(X, axis=0)
-    return torch.from_numpy(X), torch.from_numpy(Y)
 
 # CCP classes
